Add helper files to save folder. Rename raw data in default parameters.

+++ helpers/save.py
@@ -43,15 +43,30 @@ def init_save_data_structure(data_save_path,
         os.makedirs(helper_folder_name)
 
     try:
-        # put in static things
-        copy2(parameters_used, save_folder_run_name)
+        # Go to folder helpers and list files within it
+        src_files = os.listdir('helpers/')
+
+        # Loop through files in helpers and copy important ones over
+        for file_name in src_files:
+
+            # get the full name
+            full_file_name = os.path.join('helpers/', file_name)
+
+            # Check that constructed file is a real file and ends in .py
+            if (os.path.isfile(full_file_name)) and '.py' in file_name:
+                # Copy over!
+                copy2(full_file_name, helper_folder_name)
+
+        # Check that parameters file given is a real file
+        if (os.path.isfile(parameters_used)):
+            # Copy over parameters file
+            copy2(parameters_used, save_folder_run_name)
 
     # catch IO exceptions
     except IOError as error:
         if error.errno == 2:
             raise error
 
-    # return path for completion or other data type saving needs
     return save_folder_run_name
 
 
@@ -100,7 +115,7 @@ def _save_session_related_data(file, session_dictionary):
 
     # Use the file to dump data to
     try:
-        json.dump(session_dictionary, file)
+        json.dump(session_dictionary, file, indent=2)
     except Exception as e:
         raise e
